requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
cfscrape==2.1.1
charset-normalizer==3.4.0
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pyparsing==3.2.0
requests==2.32.3
requests-toolbelt==1.0.0
soupsieve==2.6
urllib3==2.2.3
Werkzeug==3.1.3
python-jobspy>=1.1.80
pydantic>=2.0.0
aiohttp>=3.9.0
